+++ beast-of-argh/CHANGELOG.md
@@ -7,6 +7,7 @@
 * Update loki and promtail to 2.7.5
 * Update loki configuration file, permissions, sysrc entries
 * Update grafana for grafana.ini and copy provisions dir from /usr/local/etc/grafana now
+* Fix permissions on grafana.ini
 
 ---